+++ clash-lib/src/app/api/mod.rs
@@ -28,6 +28,7 @@ use super::{
 };
 
 mod handlers;
+mod ipc;
 mod middlewares;
 
 pub struct AppState {
@@ -86,7 +87,7 @@ pub fn get_api_runner(
 
     let runner = async move {
         info!("Starting API server");
-        let mut app = Router::new()
+        let mut router = Router::new()
             .route("/", get(handlers::hello::handle))
             .route("/logs", get(handlers::log::handle))
             .route("/traffic", get(handlers::traffic::handle))
@@ -128,7 +129,7 @@ pub fn get_api_runner(
             .layer(ServiceBuilder::new().layer(TraceLayer::new_for_http()));
 
         if let Some(external_ui) = controller_cfg.external_ui {
-            app = app
+            router = router
                 .route("/ui", get(|| async { Redirect::to("/ui/") }))
                 .nest_service(
                     "/ui/",
@@ -145,7 +146,7 @@ pub fn get_api_runner(
             } else {
                 bind_addr
             };
-            let app_clone = app.clone();
+            let router_clone = router.clone();
             Some(async move {
                 info!("Starting API server on TCP address {bind_addr}");
                 let listener = tokio::net::TcpListener::bind(&bind_addr).await?;
@@ -188,7 +189,7 @@ pub fn get_api_runner(
                 }
                 axum::serve(
                     listener,
-                    app_clone.into_make_service_with_connect_info::<SocketAddr>(),
+                    router_clone.into_make_service_with_connect_info::<SocketAddr>(),
                 )
                 .await
                 .map_err(|x| {
@@ -200,139 +201,7 @@ pub fn get_api_runner(
             None
         };
         // Handle IPC listening
-        let ipc_fut = if let Some(ipc_path) = ipc_addr {
-            #[cfg(unix)]
-            {
-                use axum::{
-                    extract::connect_info::Connected, serve::IncomingStream,
-                };
-                use tokio::net::UnixListener;
-                let path = PathBuf::from(ipc_path);
-                let app_clone = app.clone();
-                Some(async move {
-                    info!("Start API server on IPC address {:?}", path);
-
-                    if let Err(e) = tokio::fs::remove_file(&path).await
-                        && e.kind() != std::io::ErrorKind::NotFound
-                    {
-                        return Err(crate::Error::Operation(format!(
-                            "Cannot remove existing IPC file: {e}",
-                        )));
-                    }
-
-                    if let Some(parent) = path.parent() {
-                        tokio::fs::create_dir_all(parent).await.map_err(|e| {
-                            crate::Error::Operation(format!(
-                                "Cannot create IPC dir: {e}"
-                            ))
-                        })?;
-                    }
-
-                    let uds =
-                        tokio::net::UnixListener::bind(&path).map_err(|e| {
-                            crate::Error::Operation(format!(
-                                "Cannot bind on IPC address: {e}"
-                            ))
-                        })?;
-
-                    #[derive(Clone, Debug)]
-                    #[allow(dead_code)]
-                    struct UdsConnectInfo {
-                        peer_addr: Arc<tokio::net::unix::SocketAddr>,
-                        peer_cred: tokio::net::unix::UCred,
-                    }
-
-                    impl Connected<IncomingStream<'_, UnixListener>> for UdsConnectInfo {
-                        fn connect_info(
-                            stream: IncomingStream<'_, UnixListener>,
-                        ) -> Self {
-                            let peer_addr = stream.io().peer_addr().unwrap();
-                            let peer_cred = stream.io().peer_cred().unwrap();
-                            Self {
-                                peer_addr: Arc::new(peer_addr),
-                                peer_cred,
-                            }
-                        }
-                    }
-
-                    axum::serve(
-                        uds,
-                        app_clone
-                            .into_make_service_with_connect_info::<UdsConnectInfo>(),
-                    )
-                    .await
-                    .map_err(|e| {
-                        error!("IPC API Server error: {}", e);
-                        crate::Error::Operation(format!("IPC API Server error: {e}"))
-                    })
-                })
-            }
-            #[cfg(windows)]
-            {
-                use tokio::net::windows::named_pipe::ServerOptions;
-
-                info!("Starting API server on NamedPipe {ipc_path}");
-                let pipe_name = ipc_path;
-                let app_clone = app.clone();
-                let server = ServerOptions::new()
-                    .first_pipe_instance(true)
-                    .create(&pipe_name)
-                    .map_err(|e| {
-                        crate::Error::Operation(format!("Cannot create pipe {e}"))
-                    })?;
-                Some(async move {
-                    let mut server = server;
-                    loop {
-                        server.connect().await.map_err(|e| {
-                            crate::Error::Operation(format!("NamedPipe error: {e}"))
-                        })?;
-                        let connected_client = server;
-                        server = ServerOptions::new()
-                            .first_pipe_instance(true)
-                            .create(&pipe_name)
-                            .map_err(|e| {
-                                crate::Error::Operation(format!(
-                                    "Cannot create pipe: {e}"
-                                ))
-                            })?;
-                        let app_clone = app_clone.clone();
-                        tokio::spawn(async move {
-                            use hyper_util::rt::TokioIo;
-
-                            let io = TokioIo::new(connected_client);
-                            let hyper_service = hyper::service::service_fn(
-                                move |request: hyper::Request<
-                                    hyper::body::Incoming,
-                                >| {
-                                    use tower::Service as _;
-
-                                    app_clone.clone().call(request)
-                                },
-                            );
-
-                            if let Err(e) =
-                                hyper::server::conn::http1::Builder::new()
-                                    .serve_connection(io, hyper_service)
-                                    .await
-                            {
-                                error!("NamedPipe error: {}", e);
-                            }
-                        });
-                    }
-                })
-            }
-            #[cfg(all(not(unix), not(windows)))]
-            {
-                error!("IPC only get supported on Unix and Windows");
-                Some(async move {
-                    Err(crate::Error::Operation(
-                        "IPC only get supported on Unix and Windows".to_string(),
-                    ))
-                })
-            }
-        } else {
-            None
-        };
+        let ipc_fut = ipc_addr.map(|ipc_path| async move { ipc::serve_ipc(router, &ipc_path).await });
         match (tcp_fut, ipc_fut) {
             (Some(tcp), Some(ipc)) => {
                 tokio::select! {
@@ -342,7 +211,7 @@ pub fn get_api_runner(
             }
             (Some(tcp), None) => tcp.await,
             (None, Some(ipc)) => ipc.await,
-            (None, None) => unreachable!(),
+            (None, None) => Ok(()),
         }
     };
     Some(Box::pin(runner))
